refactor(core): restructure with centralized nodes and base abstractions
