+++ src/mercurytools/core/__init__.py
@@ -1,5 +1,6 @@
-from .base import LinearBase
-from .node import Node
+from .base_linear import LinearBase
+from .base_tree import TreeBase
+from .nodes import LinearNode, LRUNode, BinaryTreeNode
 from .exceptions import *
 
-__all__=["LinearBase","Node"]
+__all__=["LinearBase","TreeBase","LinearNode","LRUNode","BinaryTreeNode"]
